ルーティングの修正、Reactでカテゴリからメニューを選べるようにする
