+++ app/assets/javascripts/components/admin_restaurant_list.js.jsx
