+++ app/assets/stylesheets/menus.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the menus controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: http://sass-lang.com/
