+++ app/controllers/menus_controller.rb
@@ -1,0 +1,5 @@
+class MenusController < ApplicationController
+  def show
+    @menu = Menu.find(params[:id])
+  end
+end
