+++ app/views/menus/show.html.erb
@@ -1,0 +1,7 @@
+<h1>メニュー詳細</h1>
+
+<%= @menu.name %>
+<%= @menu.description %>
+<%= @menu.price %>
+
+TODO: あと適当にマークアップして表示
